fix: Fix copy popup menu not dismissing when clicking elsewhere
While on that also add missing copyright headers in test files
